Removes redundant cast

+++ controllers/teleport_master_controller/teleport_master_controller.cpp
@@ -129,7 +129,7 @@ int main(int argc, char** argv) {
             // There will be a position for every robot in the positions vector
             otherRobotsNodes[i]->getField("translation")->setSFVec3f(positions[i].data());
             // Apply new rotation
-            otherRobotsNodes[i]->getField("rotation")->setSFRotation(rotations[size_t(rotDistrib(gen))].data());
+            otherRobotsNodes[i]->getField("rotation")->setSFRotation(rotations[rotDistrib(gen)].data());
 
             // Reset physics to avoid robot tearing itself apart
             otherRobotsNodes[i]->resetPhysics();
